+++ CHANGELOG.md
@@ -43,6 +43,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 private key to the cosmos keyring
 * [#44](https://github.com/babylonlabs-io/covenant-emulator/pull/44) Command
 to derive child private keys from the master key
+* [#45](https://github.com/babylonlabs-io/covenant-emulator/pull/45) Add e2e test
+with encrypted file keyring usage
 
 ## v0.9.0