+++ CHANGELOG.md
@@ -50,6 +50,7 @@ through API
 * [#48](https://github.com/babylonlabs-io/covenant-emulator/pull/48) Add covenant
 signer version that requires unlocking
 * [#49](https://github.com/babylonlabs-io/covenant-emulator/pull/49) Add show-key command
+* [#52](https://github.com/babylonlabs-io/covenant-emulator/pull/52) Fix CosmosKeyStoreConfig
 
 ## v0.9.0